+++ .idea/misc.xml
@@ -30,6 +30,8 @@
         <entry key="..\:/Users/JH/AndroidStudioProjects/Clock/app/src/main/res/layout/alarm_vibration.xml" value="0.12639311043566362" />
         <entry key="..\:/Users/JH/AndroidStudioProjects/Clock/app/src/main/res/layout/alarmitem.xml" value="0.3338541666666667" />
         <entry key="..\:/Users/JH/AndroidStudioProjects/Clock/app/src/main/res/layout/border.xml" value="0.3338541666666667" />
+        <entry key="..\:/Users/JH/AndroidStudioProjects/Clock/app/src/main/res/layout/fragment_alarm_holiday.xml" value="0.1" />
+        <entry key="..\:/Users/JH/AndroidStudioProjects/Clock/app/src/main/res/layout/fragment_global_time.xml" value="0.10507246376811594" />
         <entry key="..\:/Users/jhobo/StudioProjects/Clock/app/src/main/res/layout/activity_main.xml" value="0.3536458333333333" />
         <entry key="..\:/Users/jhobo/StudioProjects/Clock/app/src/main/res/layout/alarm_create_page.xml" value="0.3536458333333333" />
         <entry key="..\:/Users/jhobo/StudioProjects/Clock/app/src/main/res/layout/alarm_item.xml" value="0.3536458333333333" />
