+++ app/src/main/AndroidManifest.xml
@@ -19,11 +19,11 @@
             </intent-filter>
         </activity>
         <activity
-        android:name="com.eumji.clock.alarm.AlarmDetail"
+        android:name="com.eumji.clock.alarm.AlarmDetailActivity"
         android:exported="true">
         </activity>
         <activity
-            android:name="com.eumji.clock.alarm.AlarmDetailSetting"
+            android:name="com.eumji.clock.alarm.AlarmDetailSettingActivity"
             android:exported="true"
             tools:ignore="Instantiatable">
         </activity>
