+++ app/src/main/java/com/eumji/clock/MainActivity.kt
@@ -4,10 +4,10 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.View
 import android.widget.Button
-import com.eumji.clock.alarm.Alarm
-import com.eumji.clock.globaltimes.Globaltimes
-import com.eumji.clock.stopwatch.Stopwatch
-import com.eumji.clock.timer.Timer
+import com.eumji.clock.alarm.AlarmFragment
+import com.eumji.clock.globalTime.GlobalTimeFragment
+import com.eumji.clock.stopwatch.StopwatchFragment
+import com.eumji.clock.timer.TimerFragment
 
 class MainActivity : AppCompatActivity() ,View.OnClickListener {
 
@@ -15,16 +15,16 @@ class MainActivity : AppCompatActivity() ,View.OnClickListener {
         super.onCreate(savedInstanceState)
         val transaction = supportFragmentManager.beginTransaction()
         setContentView(R.layout.activity_main)
-        transaction.add(R.id.page_view, Alarm())
+        transaction.replace(R.id.page_view, AlarmFragment())
         transaction.commit()
 
         val buttonAlarm = findViewById<Button>(R.id.button_alarm)
-        val buttonGlobaltime = findViewById<Button>(R.id.button_globaltime)
+        val buttonGlobalTime = findViewById<Button>(R.id.button_global_time)
         val buttonStopwatch = findViewById<Button>(R.id.button_stopwatch)
         val buttonTimer = findViewById<Button>(R.id.button_timer)
 
         buttonAlarm.setOnClickListener(this)
-        buttonGlobaltime.setOnClickListener(this)
+        buttonGlobalTime.setOnClickListener(this)
         buttonStopwatch.setOnClickListener(this)
         buttonTimer.setOnClickListener(this)
     }
@@ -33,19 +33,19 @@ class MainActivity : AppCompatActivity() ,View.OnClickListener {
         val transaction = supportFragmentManager.beginTransaction()
         when (v.id) {
             R.id.button_alarm -> {
-                transaction.replace(R.id.page_view,Alarm())
+                transaction.replace(R.id.page_view, AlarmFragment())
                 transaction.commit()
             }
-            R.id.button_globaltime -> {
-                transaction.replace(R.id.page_view,Globaltimes())
+            R.id.button_global_time -> {
+                transaction.replace(R.id.page_view, GlobalTimeFragment())
                 transaction.commit()
             }
             R.id.button_stopwatch -> {
-                transaction.replace(R.id.page_view,Stopwatch())
+                transaction.replace(R.id.page_view, StopwatchFragment())
                 transaction.commit()
             }
             R.id.button_timer -> {
-                transaction.replace(R.id.page_view,Timer())
+                transaction.replace(R.id.page_view, TimerFragment())
                 transaction.commit()
             }
         }
