+++ app/src/main/java/com/eumji/clock/alarm/AlarmHolidayFragment.kt
@@ -7,18 +7,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.eumji.clock.R
 
-class AlarmHoliday : Fragment(){
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-        }
-    }
+class AlarmHolidayFragment : Fragment(){
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.alarm_holiday, container, false)
+        return inflater.inflate(R.layout.fragment_alarm_holiday, container, false)
     }
 }
