+++ app/src/main/java/com/eumji/clock/alarm/AlarmRecyclerAdapter.kt
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.eumji.clock.R
 import java.util.ArrayList
 
-class AlarmRecyclerAdapter(private val items: ArrayList<AlarmItem>) :
+class AlarmRecyclerAdapter(private val items: ArrayList<Alarm>) :
     RecyclerView.Adapter<AlarmRecyclerAdapter.ViewHolder>(){
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
@@ -20,7 +20,7 @@ class AlarmRecyclerAdapter(private val items: ArrayList<AlarmItem>) :
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         val listener = View.OnClickListener { _ ->
-            val intent= Intent(holder.itemView.context, AlarmDetail::class.java)
+            val intent= Intent(holder.itemView.context, AlarmDetailActivity::class.java)
             ContextCompat.startActivity(holder.itemView.context, intent, null)
         }
         holder.bind(listener, items[position])
@@ -29,7 +29,7 @@ class AlarmRecyclerAdapter(private val items: ArrayList<AlarmItem>) :
     override fun getItemCount(): Int = items.size
 
     class ViewHolder(private var v: View) : RecyclerView.ViewHolder(v) {
-        fun bind(listener: View.OnClickListener, item: AlarmItem) {
+        fun bind(listener: View.OnClickListener, item: Alarm) {
             v.setOnClickListener(listener)
         }
     }
