+++ app/src/main/java/com/eumji/clock/alarm/AlarmVibrationFragment.kt
@@ -7,17 +7,12 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.eumji.clock.R
 
-class AlarmVibration : Fragment(){
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        arguments?.let {
-        }
-    }
+class AlarmVibrationFragment : Fragment(){
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.alarm_vibration, container, false)
+        return inflater.inflate(R.layout.fragment_alarm_vibration, container, false)
     }
 }
