+++ app/src/main/java/com/eumji/clock/globalTime/GlobalTimeFragment.kt
@@ -1,4 +1,4 @@
-package com.eumji.clock.globaltimes
+package com.eumji.clock.globalTime
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -10,7 +10,7 @@ import com.eumji.clock.R
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-class Globaltimes : Fragment() {
+class GlobalTimeFragment : Fragment() {
     private var param1: String? = null
     private var param2: String? = null
 
@@ -27,13 +27,13 @@ class Globaltimes : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_globaltimes, container, false)
+        return inflater.inflate(R.layout.fragment_global_time, container, false)
     }
 
     companion object {
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            Globaltimes().apply {
+            GlobalTimeFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
