+++ app/src/main/java/com/eumji/clock/stopwatch/StopwatchFragment.kt
@@ -10,7 +10,7 @@ import com.eumji.clock.R
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-class Stopwatch : Fragment() {
+class StopwatchFragment : Fragment() {
     private var param1: String? = null
     private var param2: String? = null
 
@@ -33,7 +33,7 @@ class Stopwatch : Fragment() {
     companion object {
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            Stopwatch().apply {
+            StopwatchFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
