+++ app/src/main/java/com/eumji/clock/timer/TimerFragment.kt
@@ -10,7 +10,7 @@ import com.eumji.clock.R
 private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
-class Timer : Fragment() {
+class TimerFragment : Fragment() {
     private var param1: String? = null
     private var param2: String? = null
 
@@ -33,7 +33,7 @@ class Timer : Fragment() {
     companion object {
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            Timer().apply {
+            TimerFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
