+++ app/src/main/res/layout/activity_main.xml
@@ -32,7 +32,7 @@
             android:textColor="@color/font2"/>
 
         <Button
-            android:id="@+id/button_globaltime"
+            android:id="@+id/button_global_time"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_weight="1"
