+++ app/src/main/res/layout/fragment_alarm.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".alarm.Alarm">
+    tools:context=".alarm.AlarmFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/alarm_appbar"
