+++ app/src/main/res/layout/fragment_alarm_vibration.xml
