+++ app/src/main/res/layout/fragment_global_time.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".globaltimes.Globaltimes">
+    tools:context=".globalTime.GlobalTimeFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -25,7 +25,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="horizontal"
-            android:gravity="right">
+            android:gravity="end">
 
             <ImageButton
                 android:id="@+id/globaltimes_plus"
