+++ app/src/main/res/layout/fragment_stopwatch.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".stopwatch.Stopwatch">
+    tools:context=".stopwatch.StopwatchFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -25,7 +25,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="horizontal"
-            android:gravity="right">
+            android:gravity="end">
 
             <ImageButton
                 android:id="@+id/stopwatch_plus"
